add capability to get the valid move locations from the manager
Valid move locations depend on both the currently selected piece and the
type of manager. In a chess game, the rules of chess are enforced. In a
demo, all unoccupied squares are possible move locations. However,
highlighting this many squares on the display would be a nuissance, so
getValidMoveLocations returns an empty list in the demo mode